+++ documents/ipynb/h2ocond.rst
@@ -295,7 +295,7 @@ minimization using Primary Interior Point Method (pIPM)
 
 .. code:: ipython3
 
-    from exogibbs.optimize.pipm_cond import minimize_gibbs_cond_core
+    from exogibbs.optimize.pipm_rgie_cond import minimize_gibbs_cond_core
     import jax.numpy as jnp
     from exogibbs.api.chemistry import ThermoState
