+++ documents/ipynb/pipm/h2ocond.py
@@ -134,7 +134,7 @@ p = 1.0 # bar
 
 
 
-from exogibbs.optimize.pipm_cond import minimize_gibbs_cond_core
+from exogibbs.optimize.pipm_rgie_cond import minimize_gibbs_cond_core
 import jax.numpy as jnp
 from exogibbs.api.chemistry import ThermoState
